form to add a pet

+++ api/Data/Pet.cs
@@ -6,7 +6,7 @@ public class Pet
     public string Name { get; set; }
     public int OwnerId { get; set; }
     public User Owner { get; set; }
-    public int Age { get; set; }
+    public float Age { get; set; }
     public GenderType Gender { get; set; }
     public string Species { get; set; }
 }
